Add merge commands
