+++ src/autocomplete.js
@@ -1,0 +1,16 @@
+var isValidRequest = require('./mapzenSearch').isValidRequest;
+var sendRequest = require('./mapzenSearch').sendRequest;
+var makeResponse = require('./mapzenSearch').makeResponse;
+
+var api_key = process.env.SEARCH_API_KEY;
+
+module.exports.search = function (req, res) {
+  if (!isValidRequest(req, res)) {
+    return;
+  }
+
+  sendRequest('autocomplete', api_key, req.query.text, function (err, url, places) {
+    var response = makeResponse(url, places);
+    res.send(response);
+  });
+};
